requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr<2.0,>=1.3
Babel>=1.3
Flask>=0.10,<1.0
jsonschema!=2.5.0,<3.0.0,>=2.0.0
netaddr>=0.7.12
oslo.concurrency>=2.3.0         # Apache-2.0
oslo.serialization>=1.4.0               # Apache-2.0
oslo.utils>=2.0.0                       # Apache-2.0
python-neutronclient>=2.3.11,<3
pyroute2>=0.3.10  # Apache-2.0 (+ dual licensed GPL2)
